+++ src/main/java/Transaction.java
@@ -7,15 +7,15 @@ public class Transaction {
 
     private String typeOfOperation;
     private final String date;
-    private final int amount;
+    private final double amount;
 
-    public Transaction(String typeOfOperation, String date, int amount) {
+    public Transaction(String typeOfOperation, String date, double amount) {
         this.typeOfOperation = typeOfOperation;
         this.date = date;
         this.amount = amount;
     }
 
-    public Transaction(String date, int amount) {
+    public Transaction(String date, double amount) {
         this.date = date;
         this.amount = amount;
     }
